requirements.txt
confluent-kafka==2.0.2
PyYAML==6.0
datetime
sqlalchemy
pandas
# psycopg2
psycopg2-binary